# Update ways to evaluate kernel value when using kerevalmeth=0 (i.e. exp(sqrt())) in interpolation kernels