@ahbarnett 2022-02-08
Here are timings on an A100:

before:
2d: 1.4ms
3d: 4.8ms

after:
2d: 1.4ms
3d: 2.5ms

So, 2d no effect, 3d twice as fast.
Pretty important.
Cheers, Alex