@ahbarnett 2022-02-08
I merged, so you could do the above on another PR or just push to master. Also it would be good to get into the habit of listing improvements in the CHANGELOG...