@MelodyShih 2022-02-09
Sounds like a good idea to merge the kernel evaluation calls into one. I will experiment with it. And yes, I will remind myself to update the CHANGELOG in the future. Thank you for updating it with the recent changes.